refactor(updaters): change rootReducer into updater
Fix #4
